parse no_mangle in cfg_attr
fixes #36

the predicate currently isn't checked, which may cause false positives,
but this is very unlikely to happen and this can be dealt with in the future.
the relevant documentation and example have been adjusted.

the documentation for #[no_mangle_if_debug] has been removed since it doesn't
work with optimizations or the feature flag anyways.
it should later be deprecated and phased out.